backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
